src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
anthropic==0.7.0
python-dotenv==1.0.0
playwright==1.40.0
pillow==10.1.0
python-multipart==0.0.6
